Add files via upload
In Spring Framework, autowiring is a feature that enables you to inject the dependencies automatically. This is particularly useful in eliminating the need for explicit wiring and configuration, making the development process more streamlined and less error-prone.

Autowiring in Spring
Spring supports several types of autowiring:

byType: Autowires by property type. Spring looks for a bean with the same type as the property to be autowired.
byName: Autowires by property name. Spring looks for a bean with the same name as the property to be autowired.
constructor: Autowires by constructor. If there is exactly one bean constructor, then the dependency is autowired by the constructor.
autodetect: Spring first tries to autowire by constructor, and if it doesn't find a match, it falls back to autowiring by type.
You can enable autowiring in Spring XML configuration using the autowire attribute in the <bean> tag, or in Java configuration using the @Autowired annotation.

@Qualifier Annotation
When you have multiple beans of the same type and you want to specify which bean should be injected, you use the @Qualifier annotation along with @Autowired.